+++ MasterPassword/ObjC/MPAppDelegate_Store.m
@@ -701,7 +701,7 @@ PearlAssociatedObjectProperty(NSManagedObjectContext*, MainManagedObjectContext,
 
         [export appendFormat:@"%@  %8ld  %8s  %20s\t%@\n",
                              [[NSDateFormatter rfc3339DateFormatter] stringFromDate:lastUsed], (long)uses,
-                             [PearlString( @"%lu:%lu", type, (unsigned long)version ) UTF8String], [name UTF8String], content
+                             [PearlString( @"%lu:%lu", (long)type, (unsigned long)version ) UTF8String], [name UTF8String], content
                                                                                                                      ? content: @""];
     }
